Moved all sub-directories into root directory; added .prettierconfig file; deleted redundant node modules;
