+++ package.json
@@ -6,7 +6,8 @@
     "react": "^16.6.0",
     "react-dom": "^16.6.0",
     "react-scripts": "2.0.5",
-    "reactstrap": "^6.5.0"
+    "reactstrap": "^6.5.0",
+    "styled-components": "^4.0.2"
   },
   "scripts": {
     "start": "react-scripts start",
